Physics isn't working as well as in the JS
For some reason, the physics isn't working as well as in the JavaScript. There's still some overlapping in the final .csv (check it in the "Array Inputs.html" file. SUPER useful here. It may be in the actual physics code, or it could be in the way I had to set up the main() function. In python, we can't just "let it run until it looks good," so I tried to compensate by creating set time slots to allocate for each circle to settle, but this may need to be adapted.