Add files via upload
The files in Apollonius_graph_2 folder need to be built into a program first using CMake on a computer that has CGAL and Qt 5 installed. There is a folder (build-Apollonius_graph_2-Desktop....) that has the program already built, but it was built using my linux machine, which has different settings that may make the program un-openable on someone else's machine.